requirements.txt
asyncio==3.4.3
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
Flask==1.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.18.0
redis==3.5.3
requests==2.25.1
typing==3.7.4.1
urllib3==1.26.14
Werkzeug==1.0.1
yapf==0.32.0
